+++ src/componentes/estadistica.js
@@ -1,0 +1,12 @@
+import { computeStats } from "../dataFunctions.js";
+
+export const estadistica = (data) => {
+  const contenedorEstadistica = document.createElement("div");
+  contenedorEstadistica.setAttribute("id", "conteo");
+  const parrafoEstadística = document.createElement("p");
+  contenedorEstadistica.textContent = "Total de álbums:";
+  let sumaData = computeStats(data);
+  parrafoEstadística.innerHTML = `Total de Albums: ${sumaData}`;
+  contenedorEstadistica.appendChild(parrafoEstadística);
+  return contenedorEstadistica;
+};
