+++ src/componentes/filtros.js
@@ -38,61 +38,4 @@ export const estructuraFiltro = () => {
   return etiqueta;
 };
 
-export const estructuraOrdemaniento = () => {
-  const etiqueta = document.createElement("label");
-  etiqueta.textContent = "Ordenar:";
-  etiqueta.setAttribute("for", "ordenado");
-  const orden = document.createElement("select");
-  orden.setAttribute("data-testid", "select-sort");
-  orden.setAttribute("name", "sort-order");
-  orden.setAttribute("id", "ordenado");
 
-  const opcion1 = document.createElement("option");
-  opcion1.setAttribute("value", "blanco2");
-  opcion1.textContent = "Selecciona una opción";
-
-  const opcion2 = document.createElement("option");
-  opcion2.setAttribute("value", "asc");
-  opcion2.textContent = "A - Z";
-
-  const opcion3 = document.createElement("option");
-  opcion3.setAttribute("value", "desc");
-  opcion3.textContent = "Z - A";
-
-  orden.appendChild(opcion1);
-  orden.appendChild(opcion2);
-  orden.appendChild(opcion3);
-
-  etiqueta.appendChild(orden);
-
-  /*const contenedorOrdenamiento = document.getElementById("root");
-  contenedorOrdenamiento.addEventListener('change', function (event){
-    const ordenando = sortData(newData, "name", event.target.value);
-  PruebaHome.innerHTML = renderItems(ordenando);*/
-
-  return etiqueta;
-};
-
-export const estadistica = (data) => {
-  const contenedorEstadistica = document.createElement("div");
-  const parrafoEstadística = document.createElement("p");
-  contenedorEstadistica.textContent = "Total de álbums:";
-  let sumaData = computeStats(data);
-  parrafoEstadística.innerHTML = `Total de Albums: ${sumaData}`;
-  contenedorEstadistica.appendChild(parrafoEstadística);
-  return contenedorEstadistica;
-};
-
-//Boton
-export const boton = () => {
-  const boton = document.createElement("button");
-  boton.textContent = "Limpiar";
-  boton.id = "botonLimpiar";
-  /*boton.addEventListener("click", function () {
-    const limpiarEvent = new Event("limpiar");
-    document.dispatchEvent(limpiarEvent);
-
-    boton.appendChild(limpiarEvent);
-  });*/
-  return boton;
-};
