+++ src/index.js
@@ -21,12 +21,14 @@ import {PruebaHome} from './views/Home.js';
 // ... import other views
 import { setRootEl, setRoutes, onURLChange } from './router.js';
 import Clother from './views/Clother.js';
+import { About } from './views/About.js';
 
 // Define your routes and their associated views
 const routes = {
   "/": PruebaHome,
   "/ropa": Clother,
   "/error" : 'Hola',
+  "/about" : About,
 };
 
 const viewContainer = document.getElementById('root');
